api/eval/arcface/requirements.txt
pydantic >= 2.0.0 # required for inference.sh
inferencesh # required for inference.sh
insightface
onnxruntime-gpu>=1.18
numpy
Pillow
opencv-python-headless
